Update: item type 폴더 이동
