+++ src/common/types/item.type.ts
@@ -1,0 +1,15 @@
+export interface MediaItem {
+  id: number;
+  sns: string;
+  address: string;
+}
+
+export interface ExpertiseItemType {
+  id: number;
+  title: string;
+  sector: string;
+  industry: string;
+  businessType: string;
+  department: string;
+  role: string;
+}
