--- src/domain/signup/sections/BiographySection/types/item.type.ts
@@ -1,5 +1,0 @@
-export interface MediaItem {
-  id: number;
-  sns: string;
-  address: string;
-}
