+++ src/domain/signup/sections/ExpertiseSection/types/item.type.ts
@@ -1,13 +1,3 @@
-export interface ExpertiseItemType {
-  id: number;
-  title: string;
-  sector: string;
-  industry: string;
-  businessType: string;
-  department: string;
-  role: string;
-}
-
 export interface ExpertiseItemProps {
   sector: (id: number, newTxt: string) => void;
   industry: (id: number, newTxt: string) => void;
